WIP: refactor extensions tests
