+++ common/src/testFixtures/kotlin/org/onflow/flow/common/test/AbstractFlowEmulatorExtension.kt
@@ -1,4 +1,4 @@
-package org.onflow.flow.sdk.test
+package org.onflow.flow.common.test
 
 import org.onflow.flow.sdk.*
 import org.onflow.flow.sdk.cadence.StringField
