+++ java-example/src/test/java/org/onflow/examples/java/AccessAPIConnectorTest.java
@@ -3,6 +3,7 @@ package org.onflow.examples.java;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import org.onflow.flow.common.test.*;
 import org.onflow.flow.sdk.FlowAccessApi;
 import org.onflow.flow.sdk.FlowAddress;
 import org.onflow.flow.sdk.crypto.Crypto;
