+++ sdk/src/intTest/org/onflow/flow/sdk/transaction/TransactionIntegrationTest.kt
@@ -1,7 +1,7 @@
 package org.onflow.flow.sdk.transaction
 
 import org.onflow.flow.sdk.*
-import org.onflow.flow.sdk.test.FlowEmulatorTest
+import org.onflow.flow.common.test.FlowEmulatorTest
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.fail
