+++ sdk/src/test/kotlin/org/onflow/flow/sdk/ScriptTest.kt
@@ -2,9 +2,9 @@ package org.onflow.flow.sdk
 
 import org.onflow.flow.sdk.cadence.*
 import org.onflow.flow.sdk.crypto.Crypto
-import org.onflow.flow.sdk.test.FlowEmulatorProjectTest
-import org.onflow.flow.sdk.test.FlowTestClient
-import org.onflow.flow.sdk.test.FlowTestUtil
+import org.onflow.flow.common.test.FlowEmulatorProjectTest
+import org.onflow.flow.common.test.FlowTestClient
+import org.onflow.flow.common.test.FlowTestUtil
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Assertions.assertTrue
 import org.junit.jupiter.api.Test
